pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "guhcampos"
version = "0.1.0"
description = "Build tools for guhcampos.github.io"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Gustavo Campos", email = "guhcampos@gmail.com"},
]
keywords = ["hugo", "website", "build", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pathspec>=0.11.0",
    "watchdog>=3.0.0",
    "requests>=2.31.0",
    "pydantic-settings>=2.10.1",
    "python-slugify>=8.0.4",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
guhcampos = "guhcampos.cli:guhcampos"

[project.urls]
Homepage = "https://github.com/guhcampos/guhcampos.github.io"
Repository = "https://github.com/guhcampos/guhcampos.github.io"

[tool.hatch.build.targets.wheel]
packages = ["src/guhcampos"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["guhcampos"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
